Add order ownership checks, optional envelope signing, CMS transactions, and protocol updates
Security:
- Add dispatch ownership verification (cancel/receipt/redirect reject
  messages from non-owning stations; core role bypasses check)
- Add optional HMAC-SHA256 envelope signing (signing_key in messaging
  config; sign on publish, verify on ingest; disabled by default)

Protocol:
- Add signing.go with Sign/VerifyAndUnwrap for envelope authentication
- Ingestor verifies signatures when SigningKey is set
- Update protocol payloads and types for blueprint catalog support

Core:
- Add CMS transactions (engine, store, handlers)
- Expand corrections engine and store
- Add demand management updates
- Update node handling with station associations
- Migrate schemas (SQLite + Postgres) for new tables
- Update UI templates (nodes, payloads, diagnostics, demand)

Edge:
- Rename style_catalog to blueprint_catalog
- Update order manager, edge handler, heartbeat for protocol changes
- Update store schemas and payload handling
- Update UI templates (material, manual-order, setup)

+++ shingo-core/store/blueprint_manifest.go
@@ -6,7 +6,7 @@ type BlueprintManifestItem struct {
 	ID          int64     `json:"id"`
 	BlueprintID int64     `json:"blueprint_id"`
 	PartNumber  string    `json:"part_number"`
-	Quantity    float64   `json:"quantity"`
+	Quantity    int64     `json:"quantity"`
 	Description string    `json:"description"`
 	CreatedAt   time.Time `json:"created_at"`
 }
